Added support for different color bikes.
tron16x16_gs_ro.png
-------------------

Added wall sprites for the different color bots and rearranged them so
that the walls are immediately after the bot sprites.

game.py
-------

Added the characters for the various bots and created an array of bot colors.

Modified create_new_player and place_bike so that place_bike takes a
parameter based on the index of the player in the self.players array.

The sprite (character) selected for the Bike() is based on the index of
the player into the bot sprites array.

There are actually 5 bot colors although right now there are only four
players. If there are additional players added, the bot colors will wrap
around and repeat.

Bikes.py
--------

The self.char for the Bike is still the parameter provided at create time.

But now, self.CHAR_START for the Bike is based on the char for the Bike
-- since the walls appear after the bot sprite, self.CHAR_START is set
using ord(self.char) + 1.

Formatted some comments. Comments should have a space following the #
mark and be line-wrapped, in my opinion.